LOG-3210: Refactor cluster rbac into recocile package
